Se agrego archivo para desinstalaer el programa y sus archivos de configuracion

+++ install.sh
@@ -2,6 +2,7 @@
 
 DIR="$(dirname "$0")"
 BIN="$DIR/bin/cerrusv"
+CFG_FILE="cerrusv.cfg"
 
 if [ ! -f "$BIN" ]; then
     echo "No se encontró el ejecutable en: $BIN"
@@ -12,7 +13,13 @@ echo "Instalando ejecutable"
 cp "$BIN" /usr/local/bin/
 chmod +x /usr/local/bin/cerrusv
 
-echo "Copiando archivo de configuracion"
-cp cerrusv.cfg /etc/
+if [ ! -f "$CFG_FILE" ]; then
+    echo "No existe el archvo de configuracion"
+else
+    echo "Copiando archivo de configuracion"
+    cp cerrusv.cfg /etc/
+fi
+
+
 
 echo "Instalado en /usr/local/bin/cerrusv"
